+++ REST_API.py
@@ -14,8 +14,10 @@ def index(template='index.html'):
 @app.route('/api/<data>', methods=['POST'])
 def api_post(data):
     """
-    When the request method is an HTTP/POST, the logic is to grab the data
-    Check if it's in the DB, if it's not, we append it to the DB
+    +-------------------------------------------------------------------------------------------------------------------
+    | When the request method is an HTTP/POST, the logic is to grab the data
+    | Check if it's in the DB, if it's not, we append it to the DB
+    +-------------------------------------------------------------------------------------------------------------------
     """
     input_params = request.get_json()
     if len(input_params) is not (DB_SIZE - 1):
@@ -31,7 +33,9 @@ def api_post(data):
 @app.route('/api/<data>', methods=['GET'])
 def api_get(data):
     """
-    If the content exist in the DB, fetch it from the DB and return to user, else reject (404)
+    +-------------------------------------------------------------------------------------------------------------------
+    | If the content exist in the DB, fetch it from the DB and return to user, else reject (404)
+    +-------------------------------------------------------------------------------------------------------------------
     """
     if is_in_db(data):
         return fetch_from_db(key=data, request_dict=request.args.to_dict())
@@ -42,11 +46,12 @@ def api_get(data):
 
 def append_to_db(key=None, request_dict=None):
     """
-    The variable content starts as None which is the recommended way according to google code style.
-    if content is None, a log message is printed with the content and type(content)
-    The parameters db and db_mode are initialized to database.csv and a (append) respectively,
-    this means this function is universal and the db/db_mode can be changed when calling the function
-    Example: append_to_db(content='some content: some value', db='new_db.csv', db_mode='w')
+    +-------------------------------------------------------------------------------------------------------------------
+    | The variable content starts as None which is the recommended way according to google code style.
+    | if content is None, a log message is printed with the content and type(content)
+    | Vendor received from URL and other parameters received by JSON
+    | Function gathers all the parameters for Car and inserts row into DB
+    +-------------------------------------------------------------------------------------------------------------------
     """
     app.logger.info("Inside the function: append_to_db()")
 
@@ -57,7 +62,8 @@ def append_to_db(key=None, request_dict=None):
         db = open_db_connection()
         cursor = db.cursor()
 
-        param = f' \'{key}\' '
+        """ Variable "param" gathers all DATA that we want to insert into DB """
+        param = f"'{key}'"
         for k, v in request_dict.items():
             param += f",'{v}'"
 
@@ -65,17 +71,22 @@ def append_to_db(key=None, request_dict=None):
         query = f"INSERT INTO cars(Vendor, Model, Year, Engine, HP, Torque)\
                         VALUES ({param});"
 
-        """Adding content """
+        """ Adding content """
         cursor.execute(query)
 
-        """Closing DB"""
+        """ Writing into DB """
         db.commit()
-
         app.logger.info(f"Successfully appended content={param}, type:{type(param)}")
+
         return f"Successfully appended content={param}, type:{type(param)}"
 
 
 def content_in_db(key=None, request_dict=None):
+    """
+     +------------------------------------------------------------------------------------------------------------------
+     | This function verifies if the specific car spec. already inside DB before we want to insert it into DB
+     +------------------------------------------------------------------------------------------------------------------
+     """
     app.logger.info("Inside the function: content_in_db()")
 
     if key is not None:
@@ -96,7 +107,7 @@ def content_in_db(key=None, request_dict=None):
             db_data = {
                 'model': str(Model), 'year': str(Year), 'engine': str(Engine), 'HP': str(HP), 'Torque': str(Torque)
             }
-
+            """ First match row in DB we return that entry exist in DB """
             if db_data == request_dict:
                 return True
             else:
@@ -107,6 +118,11 @@ def content_in_db(key=None, request_dict=None):
 
 
 def is_in_db(content=None):
+    """
+     +------------------------------------------------------------------------------------------------------------------
+     | This function verifies if the specific Vendor name exist inside DB before that we run complicated logic
+     +------------------------------------------------------------------------------------------------------------------
+     """
     app.logger.info("Inside the function: is_in_db()")
 
     db = open_db_connection()
@@ -118,10 +134,9 @@ def is_in_db(content=None):
 
 def fetch_from_db(key=None, request_dict=None):
     """
-    We pass in a key which starts out as None for error handling
-    the key is the Vendor of the car, we then open the DB
-    and check if the key matches the vendor column, if it does
-    we format the string and return it to the function which called
+    +-------------------------------------------------------------------------------------------------------------------
+    | Function searches for the entry that we want to GET, when match - we format and return
+    +-------------------------------------------------------------------------------------------------------------------
     """
     app.logger.info("Inside the function: fetch_from_db()")
 
@@ -138,12 +153,12 @@ def fetch_from_db(key=None, request_dict=None):
             params_list.append(result)
         params = " AND ".join(params_list)
 
+        """ Query can be done on short URL or URL with args """
         if not request_dict:
             query = f"SELECT * FROM cars WHERE Vendor == '{key}';"
         else:
             query = f"SELECT * FROM cars WHERE Vendor == '{key}' AND {params};"
         app.logger.debug(query)
-
         cursor = db.cursor()
 
         output = []
@@ -162,24 +177,29 @@ def fetch_from_db(key=None, request_dict=None):
         return jsonify(output)
 
     else:
-        """
-        If we didn't get a key, reject the request with a 404 not found
-        """
+        """ If we didn't get a key, reject the request with a 404 not found """
         reject_invalid_request()
 
 
 def reject_invalid_request(code=404):
     """
-    This function is designed to reduce the lines of logic in the routes by 
-    Having a generic method to abort requests, when needed, just call reject_invalid_request
-    A log message will be printed and the request rejected with a 404 not found by default
-    Can be changed by calling the function with a different code example: reject_invalid_request(500)
+    +-------------------------------------------------------------------------------------------------------------------
+    | This function is designed to reduce the lines of logic in the routes by
+    | Having a generic method to abort requests, when needed, just call reject_invalid_request
+    | A log message will be printed and the request rejected with a 404 not found by default
+    | Can be changed by calling the function with a different code example: reject_invalid_request(500)
+    +-------------------------------------------------------------------------------------------------------------------
     """
     app.logger.info(f"Invalid Request:{request} - Dropping with code: {code}")
     abort(code)
 
 
 def open_db_connection(db='cars.db'):
+    """
+    +-------------------------------------------------------------------------------------------------------------------
+    | This function in late release will be removed and SQLITE3 connection will be available from main code
+    +-------------------------------------------------------------------------------------------------------------------
+    """
     conn = sqlite3.connect(db)
     return conn